Rename scopes attribute to scope
The main reason for this is the conflict with Mongoid scopes where a
kind of monkey patching was necessary in order to make this work.

This also makes the scope attribute match with the parameter that is
received on token request. A `oauth_scope` method was created when is
necessary to retrieve the scopes as an set